Add email notifications to comments
A new event has been created called CommentCreated which will hold the
information related to the event; i.e. the concern details and the
comment details. The CommentController @show method has been modified to emit this event when a new comment is created.

A new event listener has been set up called NotifyConcernAuthor which will listen for the CommentCreated event and dispatch an email notification to the User who initially created the concern.