Fix for body map saving incorrectly
Fixes an issue where creating a new concern would save the body map link
when no body map was created. This was due to $request->has('image')
being called in the @store method.

This has been updated to if ($request->image) to now check whether there
is any data in the image input field and calling the saveBodyMap()
method accordingly.