Add updateResolvedStatus method to Concern model
Adds an updateResolvedStatus() method to the Concern model. Checks if
the current status is not null and updates with the current date,
otherwise returns null if the current status is not null.

Tidies the update() method on the ConcernController by removing if
statement and making a call to the method on the model.