Implement AssemblyServiceProvider, bind MISInterface
Implements a new AssemblyServiceProvider which binds an MISInterface to
the service container. Returns a new instance of Assembly service class
with the client ID and client secret.

Updates the Assembly class to accept the client ID and secret through
the constructor and use these keys in the class.

Updates any implementation of the Assembly class throughout the code
base, i.e. in the jobs to use the MISInterface instead. This will allow
for switching out the Assembly class for another implementation if
required further down the line.